make multihash pluggable
Note: this is a suggestion/WIP which can be re-worked as needed.

This adds a map var which maps codecs to hash functions. The hash funcs must
have signature which accepts one byte slice, and returns one byte slice.

It adds a RegisterHashFunc method which allows registration for any valid
codec. It prevents double registration or overwriting of hash funcs already
registered.

It also creates two funcs which are called by init() to register the funcs
which is already the current behaviour of the code:

  * registerStdlibHashFuncs
  * registerNonStdlibHashFuncs

If desired, the non-stdlib funcs can then easily be removed. Or it can be
implemented as-is, and required that future hash funcs use the Registration
func. Or refactored in some other way if necessary.

I have tested this and it is working with IPLD Dash resolver. Can also add
example in the README if this is approved / agreed.