Resolving review topics around Encode function.
Leaving the error return in, per
https://github.com/multiformats/go-multihash/pull/136#discussion_r587708992

May also be interesting to note: I did check if ValidCode (or, roughly
the same thing but perhaps renamed to KnownCode) could be reintroduced,
and the answer is... actually no, not without broaching other issues.
A body of `_, ok := registry[code]` is enough to open a can of worms.
See
https://github.com/multiformats/go-multihash/pull/136#discussion_r589712826