Replace need for DefaultLength in opts processing.
This creates a hasher long enough to ask it its properties.
This is arguably creating garbage; on the other hand, I don't think
this is a codepath ever likely to be used in a hot loop anywhere.
We can extract this to something that caches the properties later
if it proves necessary.

It quietly defaults to zero for unknown codes.  I don't know if this
makes sense, but it's what the old code would have done, so it's what
the new code will do, and I'm not looking deeper into it.
At this point I'm just trying to make surgically minimal alterations
and get this changeset as a whole wrapped up so things can move on.