Proactively reject registration of nil functions.
Surely no one would try to do this, nor then be surprised if it
creates problems.  On the other hand: if someone *does* do this, and
the error doesn't appear until arbitrarily far away when the map is
read, it's a pain to diagnose... and checking it up front is cheap.
So, here we go: check it up front.