Improve situation around ErrSumNotSupported.
The number that's missing is now reported in the error.
This is done using the golang error wrap feature.

We're keeping a constant value for ErrSumNotSupported for
compatibility and change avoidance reasons.

Code that cares about this exact error will still have to change;
it is now necessary to use `errors.Is` to detect it.

The text in the ErrSumNotSupported value is also updated,
because the text no longer made sense given an open registry system.
As a target-of-opportunity fix, it also now follows golang normative
conventions for error messages (no capitalization, no punctuation).