Reintroduce DefaultLengths; populate during Register.
While it was possible to remove all use of this from this repo, when
attempting propagate changes to downstreams consuming it, it became
apparently that other repos also rely on this symbol.

Whether or not those usages are important and intentional, whether
they're actually worth maintaining, and whether they'd be replacable
with other approaches... is not considered at this time.
(Probably we should be asking this!  The first occasions where this
cropped up are in other functions that have been marked "deprecated"
since... 2018!  But... chasing those things down and straightening
them out is becoming problematic.  Perhaps we'll be more ready to
revisit these things at a later date.)