don't use pointers for Multihash.String
The String and HexString methods on Multihash are read-only,
so strictly speaking they do not require a pointer receiver.
The multihash type itself is a slice,
which is only three times as big as a pointer,
so a pointer isn't needed to avoid copying memory either.

However, there's a good reason to avoid a pointer receiver:
it allows method calls where an expression isn't addressable.
For example:

	$ cat input.go
	[...]
	s := block.Cid().Hash().HexString()
	[...]
	$ go build input.go
	input: cannot call pointer method HexString on "github.com/multiformats/go-multihash".Multihash

One can split the code to make the multihash addressable:

	mh := block.Cid().Hash()
	s := mh.HexString() // implicitly: (&mh).HexString()

However, that seems unnecessary when HexString does not need a pointer.